requirements.txt
absl-py>=0.10
pyyaml>=5
numpy>=1.15
pandas>=1
pytest>=6